backend/requirements.txt
Django~=5.2.0
psycopg2~=2.9.10
djangorestframework~=3.16.0
